made configurator partial
