+++ Domain.Interfaces/Repositories/IRepository.cs
@@ -1,11 +1,14 @@
 ﻿using System.Collections.Generic;
 using System;
+using RelationApp.Domain.Models;
 
 namespace RelationApp.Domain.Iterfaces
 {
     public interface IRepository<TEntity>
     {
         IEnumerable<TEntity> GetAll();
+
+        IEnumerable<TEntity> GetFiltered(Func<TEntity, bool> predicate);
 
         TEntity GetOne(Guid id);
