+++ Persistence/Configurations/RelationAppContextConfigurator.Country.cs
@@ -1,0 +1,54 @@
+﻿using Microsoft.EntityFrameworkCore;
+using RelationApp.Domain.Models;
+
+namespace Persistence.Configurations
+{
+    internal partial class RelationAppContextConfigurator
+    {
+        public static void ConfigureCountry(ModelBuilder builder)
+        {
+            builder.Entity<Country>(entity =>
+            {
+                entity.HasNoKey();
+
+                entity.ToTable("tblCountry");
+
+                entity.Property(e => e.CreatedAt).HasColumnType("datetime");
+
+                entity.Property(e => e.CreatedBy)
+                    .IsRequired()
+                    .HasMaxLength(50)
+                    .IsUnicode(false);
+
+                entity.Property(e => e.Description)
+                    .HasMaxLength(255)
+                    .IsUnicode(false);
+
+                entity.Property(e => e.Iso31662)
+                    .HasColumnName("ISO3166_2")
+                    .HasMaxLength(2)
+                    .IsUnicode(false);
+
+                entity.Property(e => e.Iso31663)
+                    .HasColumnName("ISO3166_3")
+                    .HasMaxLength(3)
+                    .IsUnicode(false);
+
+                entity.Property(e => e.ModifiedAt).HasColumnType("datetime");
+
+                entity.Property(e => e.ModifiedBy)
+                    .HasMaxLength(50)
+                    .IsUnicode(false);
+
+                entity.Property(e => e.Name)
+                    .IsRequired()
+                    .HasMaxLength(255)
+                    .IsUnicode(false);
+
+                entity.Property(e => e.PostalCodeFormat)
+                    .HasMaxLength(255)
+                    .IsUnicode(false);
+            });
+        }
+    }
+}
