+++ Persistence/Configurations/RelationAppContextConfigurator.RelationCategory.cs
@@ -1,0 +1,18 @@
+﻿using Microsoft.EntityFrameworkCore;
+using RelationApp.Domain.Models;
+
+namespace Persistence.Configurations
+{
+    internal partial class RelationAppContextConfigurator
+    {
+        public static void CreateRelationCategory(ModelBuilder builder)
+        {
+            builder.Entity<RelationCategory>(entity =>
+            {
+                entity.HasNoKey();
+
+                entity.ToTable("tblRelationCategory");
+            });
+        }
+    }
+}
