+++ Persistence/Persistence.csproj
@@ -5,6 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="7.0.0" />
     <PackageReference Include="microsoft.entityframeworkcore.sqlserver" Version="3.1.0" />
   </ItemGroup>
