+++ Relations/Views/_ViewImports.cshtml
@@ -1,4 +1,4 @@
 ﻿@using RelationApp
-@using RelationApp.Domain.Models
+@using RelationApp.Models
 @using RelationApp.Domain.Iterfaces
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
